[refactor] 파일 읽는 함수 이동
