+++ libs/components-admin/src/lib/AdminOverlayImageUploadDialog.tsx
@@ -19,12 +19,9 @@ import {
 import { ChakraNextImage } from '@project-lc/components-core/ChakraNextImage';
 import { ConfirmDialog } from '@project-lc/components-core/ConfirmDialog';
 import { ImageInput } from '@project-lc/components-core/ImageInput';
-import {
-  GoodsPreviewItem,
-  Preview,
-  readAsDataURL,
-} from '@project-lc/components-seller/GoodsRegistPictures';
+import { GoodsPreviewItem } from '@project-lc/components-seller/GoodsRegistPictures';
 import { OverlayImageTypes } from '@project-lc/shared-types';
+import { Preview, readAsDataURL } from '@project-lc/utils';
 import { s3 } from '@project-lc/utils-s3';
 import { useEffect, useState } from 'react';
